dashboard_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiokafka
kafka-python
pydantic
python-multipart
websockets
